+++ src/bbochat/forms/frm_main.py
@@ -5,7 +5,7 @@ import random
 import re
 import tkinter as tk
 from pathlib import Path
-from tkinter import simpledialog, ttk
+from tkinter import messagebox, simpledialog, ttk
 
 import clipboard
 import emoji
@@ -491,6 +491,13 @@ class MainFrame:
         self.username_2.set(data_store.username_2)
 
     def _dismiss(self, *args) -> None:
+        if self.tournament_tab.notes_changed():
+            response = tk.messagebox.askyesno(
+                "Save Notes",
+                "Do you want to save the board notes?",
+            )
+            if response:
+                self.tournament_tab.save_notes()
         self._save_sashes()
         self.root.destroy()
         # Need to do this because window_resize is called in close