+++ src/bbochat/module_caller.py
@@ -1,32 +1,17 @@
+from psiutils.module_caller import ModuleCaller as ModuleCallerBase
+
 from bbochat.data_store import data_store
 from bbochat.forms.frm_config import ConfigFrame
 from bbochat.forms.frm_edit_select import EditSelectFrame
 
 
-class ModuleCaller:
-    def __init__(self, root, module) -> None:
-        modules = {
-            "config": self._config,
-            "edit": self._edit,
+class ModuleCaller(ModuleCallerBase):
+    def __init__(self, root, parsed_args: dict) -> None:
+        self.modules = {
+            "config": (self._config, None),
+            "edit": (self._edit, "Edit greetings"),
         }
-
-        self.invalid = False
-        if module == "-h":
-            for key in sorted(list(modules.keys()) + ["main"]):
-                print(key)
-            self.invalid = True
-            return
-
-        if module not in modules:
-            if module != "main":
-                print(f"Invalid function name: {module}")
-            self.invalid = True
-            return
-
-        self.root = root
-        modules[module]()
-        self.root.destroy()
-        return
+        super().__init__(root, parsed_args)
 
     def _config(self) -> None:
         dlg = ConfigFrame(self)
@@ -35,5 +20,5 @@ class ModuleCaller:
     def _edit(self) -> None:
         ds = data_store
         ds.read()
-        dlg = EditSelectFrame(self, "greetings")
+        dlg = EditSelectFrame(self)
         self.root.wait_window(dlg.root)